refactor(nodes): Move DifyNodeFactory to core.app.workflow
