+++ api/core/app/workflow/__init__.py
@@ -1,0 +1,3 @@
+from .node_factory import DifyNodeFactory
+
+__all__ = ["DifyNodeFactory"]
